Move labReportFilter
