+++ apps/modernization-api/src/main/java/gov/cdc/nbs/controller/ExportController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import com.google.common.net.HttpHeaders;
 import com.itextpdf.text.DocumentException;
-import gov.cdc.nbs.graphql.filter.LabReportFilter;
 import gov.cdc.nbs.investigation.InvestigationFilter;
 import gov.cdc.nbs.investigation.InvestigationFinder;
+import gov.cdc.nbs.labreport.LabReportFilter;
 import gov.cdc.nbs.labreport.LabReportFinder;
 import gov.cdc.nbs.service.ExportService;
 import io.swagger.annotations.ApiImplicitParam;
