+++ apps/modernization-api/src/test/java/gov/cdc/nbs/LabReportSearchSteps.java
@@ -14,17 +14,17 @@ import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.transaction.annotation.Transactional;
 import gov.cdc.nbs.entity.elasticsearch.LabReport;
-import gov.cdc.nbs.graphql.filter.LabReportFilter;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.EntryMethod;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.EventStatus;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.LabReportDateType;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.LabReportProviderSearch;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.LaboratoryEventDateSearch;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.LaboratoryEventIdType;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.ProcessingStatus;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.ProviderType;
-import gov.cdc.nbs.graphql.filter.LabReportFilter.UserType;
+import gov.cdc.nbs.labreport.LabReportFilter;
 import gov.cdc.nbs.labreport.LabReportFinder;
+import gov.cdc.nbs.labreport.LabReportFilter.EntryMethod;
+import gov.cdc.nbs.labreport.LabReportFilter.EventStatus;
+import gov.cdc.nbs.labreport.LabReportFilter.LabReportDateType;
+import gov.cdc.nbs.labreport.LabReportFilter.LabReportProviderSearch;
+import gov.cdc.nbs.labreport.LabReportFilter.LaboratoryEventDateSearch;
+import gov.cdc.nbs.labreport.LabReportFilter.LaboratoryEventIdType;
+import gov.cdc.nbs.labreport.LabReportFilter.ProcessingStatus;
+import gov.cdc.nbs.labreport.LabReportFilter.ProviderType;
+import gov.cdc.nbs.labreport.LabReportFilter.UserType;
 import gov.cdc.nbs.message.enums.PregnancyStatus;
 import gov.cdc.nbs.repository.JurisdictionCodeRepository;
 import gov.cdc.nbs.repository.elasticsearch.LabReportRepository;
