+++ apps/modernization-api/src/test/java/gov/cdc/nbs/PermissionSteps.java
@@ -23,11 +23,11 @@ import gov.cdc.nbs.config.security.SecurityProperties;
 import gov.cdc.nbs.controller.PatientController;
 import gov.cdc.nbs.entity.enums.RecordStatus;
 import gov.cdc.nbs.graphql.GraphQLPage;
-import gov.cdc.nbs.graphql.filter.LabReportFilter;
 import gov.cdc.nbs.graphql.filter.OrganizationFilter;
 import gov.cdc.nbs.graphql.filter.PatientFilter;
 import gov.cdc.nbs.investigation.InvestigationFilter;
 import gov.cdc.nbs.investigation.InvestigationResolver;
+import gov.cdc.nbs.labreport.LabReportFilter;
 import gov.cdc.nbs.labreport.LabReportResolver;
 import gov.cdc.nbs.repository.ProgramAreaCodeRepository;
 import io.cucumber.java.Before;
